fixing opperators do operate over float arrays

+++ src/cube/aggregation/operators/AggregationOperator.java
@@ -1,5 +1,5 @@
 package cube.aggregation.operators;
 
 public interface AggregationOperator {
-	public double compute(Object args);
+	public float compute(Object args);
 }
